javaspring-buoi38

+++ src/main/java/vn/scrip/buoi38_bvn/repository/BorrowRepository.java
@@ -1,9 +1,7 @@
 package vn.scrip.buoi38_bvn.repository;
 
-import org.springframework.data.jpa.repository.JpaRepository;
 import vn.scrip.buoi38_bvn.entites.Borrow;
+import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BorrowRepository extends JpaRepository<Borrow, Long> {
-
-}
+public interface BorrowRepository extends JpaRepository<Borrow, Integer> {}
